fix: not having an mqtt server running will no longer break the runtime
Previously, the board would stop creating totp codes, and eventually hang,
if the mqtt server was suddently stopped or if there wasn't a running mqtt server to connect.
This happens when the connection attempt timeout is longer than the MQTT_RECONNECT_INTERVAL.
Now, connections aren't going to be attempted if MQTT_SERVER and MQTT_PORT config properties
aren't set. Moreoever, if the MQTT server stops, the board no longer hangs in an infinite loop.