fix: pin validation no longer throws a runtime exception
The reason it was happening is because the pinHash and pinKey were
looking at mem positions that were being constantly changed. Creating
a copy of security.pin.hash and security.pin.key solved the problem.
I believe the issue is caused by YAMLNode::gettext. I will fix this later
as my plan is to have a global Configuration object which other modules
can access when necessary.