refactor(site): remove bootstrap 5 from the build
Boostrap 5 was removed from the build to reduce the bundle size.
The size of the bundle is important because this app will
also be used as the UI for managing TOTPs in the board. Something
similar to the UI found in routers when navigating to localhost.
Moreover, the site has to be kept under 1Mb to allow the spiffs.bin
to be created and flashed in the esp32.