fix(manager): re-auth during hard refreshes while session cookie
When a User hard/soft refresh the page, a new page would be handed
over to the client, which would then force the store to reinitialize.
Because of that, the isAuthenticated flag would be empty, therefore
prompting the User to re-authenticate.
After persisting the store in the session storage, this issue no longer
occurs.

There is no security issues because tthe session cookie is still being
validated by the server or every route change, or api request.