feat: there can exist up to 100 services divided between 10 groups
instead of having a fixed amount of services per group, the services
db was enhanced to allow a dynamic number of services per group
as long as the total number of services does not surpass 100
and the number of groups does not surpass 10.

When a services.yml file contains more than 10 groups, the last ones
will be ignored.

For example, if groups 0 to 9 were already registered, a service using
group 10 will be ignored and not loaded even if there is still service
slots available.